requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
Django==5.2.6
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
redis==6.4.0
sqlparse==0.5.3
typing_extensions==4.15.0